requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
jinja2==3.1.3
motor==3.3.2
beanie==1.23.2
pymongo==4.6.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic[email]==2.7.0
google-generativeai==0.5.4
google-generativeai==0.5.4
feedparser==6.0.11
lxml==5.1.1
apscheduler==3.10.4
beautifulsoup4==4.12.3
pydantic-settings
